+++ src/main/java/seedu/duke/command/Command.java
@@ -1,9 +1,9 @@
 package seedu.duke.command;
 
 import seedu.duke.Biometrics;
-import seedu.duke.ExerciseList;
 import seedu.duke.Ui;
 import seedu.duke.exception.IllegalValueException;
+import seedu.duke.exercise.ExerciseList;
 
 public abstract class Command {
     public Command() {
