+++ src/main/java/seedu/duke/command/ViewCommand.java
@@ -1,7 +1,11 @@
 package seedu.duke.command;
 
-import seedu.duke.*;
+import seedu.duke.Biometrics;
+import seedu.duke.Parser;
+import seedu.duke.Ui;
 import seedu.duke.exception.IllegalValueException;
+import seedu.duke.exercise.Exercise;
+import seedu.duke.exercise.ExerciseList;
 
 import java.util.ArrayList;
