+++ src/main/java/seedu/duke/exercise/Exercise.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exercise;
 
 public class Exercise {
     private String exerciseName;
